reorganizing a few things
