+++ src/examples/cartridge/shopping/suppliers/mockaroo/io/index.ts
