+++ src/examples/cartridge/shopping/suppliers/mockaroo/io/mockarooHTTPClient.ts
