+++ src/examples/cartridge/shopping/suppliers/mockaroo/offersOperation/__fixtures__/PET_SHOP_OFFERS_RESPONSE.ts
