+++ src/examples/cartridge/shopping/suppliers/mockaroo/offersOperation/index.ts
