+++ src/examples/cartridge/shopping/suppliers/mockaroo/offersOperation/transformOffersOperationRequest.test.ts
