+++ src/examples/cartridge/shopping/suppliers/mockaroo/offersOperation/transformOffersOperationResponse.test.ts
