+++ src/examples/cartridge/shopping/suppliers/mockaroo/purchaseOperation/__fixtures__/PET_SHOP_PURCHASE_RESPONSE.json
