+++ src/examples/cartridge/shopping/suppliers/mockaroo/purchaseOperation/__fixtures__/PET_SHOP_PURCHASE_RESPONSE.ts
@@ -1,6 +1,6 @@
 import { map, reduce } from 'lodash';
 
-import { PetShopInventoryItem } from '@console-cartridge-contract/examples/xmlAPICartridge/types/data';
+import { PetShopInventoryItem } from '@console-cartridge-contract/examples/cartridge/shopping/types/data';
 
 import { dogTuxedoFixture, fancyCatCollarFixture } from '.';
