+++ src/examples/cartridge/shopping/suppliers/mockaroo/util/stringifyAllValuesInObject.ts
