+++ src/examples/cartridge/shopping/types/PetShoppingSupplier.ts
