+++ src/examples/cartridge/shopping/types/data/PetShopCategory.ts
