+++ src/examples/cartridge/shopping/types/data/PetShopInventoryItem.ts
