+++ src/examples/cartridge/shopping/types/data/PetShopPurchaseRequest.ts
