+++ src/examples/cartridge/shopping/types/data/PetShopPurchaseResponse.ts
