+++ src/examples/cartridge/shopping/types/operations/OffersOperation.ts
