+++ src/examples/cartridge/shopping/types/operations/PurchaseOperation.ts
